Rename math to jsPolyfills
