+++ packages/typegpu/src/tgsl/consoleLog/types.ts
@@ -2,7 +2,7 @@ import type { TgpuMutable } from '../../core/buffer/bufferShorthand.ts';
 import type { Snippet } from '../../data/snippet.ts';
 import type { AnyWgslData, Atomic, U32, WgslArray, WgslStruct } from '../../data/wgslTypes.ts';
 import type { GenerationCtx } from '../generationHelpers.ts';
-import type { supportedLogOps } from '../math.ts';
+import type { supportedLogOps } from '../jsPolyfills.ts';
 
 export type SupportedLogOp = (typeof supportedLogOps)[number];
