+++ packages/typegpu/src/tgsl/jsPolyfills.ts
